ocvalidate: Split to several sources
