+++ Utilities/ocvalidate/Makefile
@@ -6,6 +6,15 @@
 PROJECT = ocvalidate
 PRODUCT = $(PROJECT)$(SUFFIX)
 OBJS    = $(PROJECT).o \
+		  OcValidateLib.o \
+		  ValidateAcpi.o \
+		  ValidateBooter.o \
+		  ValidateDeviceProperties.o \
+		  ValidateKernel.o \
+		  ValidateMisc.o \
+		  ValidateNVRAM.o \
+		  ValidatePlatformInfo.o \
+		  ValidateUEFI.o \
           OcConfigurationLib.o
 VPATH   = ../../Library/OcConfigurationLib
 include ../../User/Makefile
