+++ Utilities/ocvalidate/ValidateAcpi.c
@@ -1,0 +1,169 @@
+/** @file
+  Copyright (C) 2018, vit9696. All rights reserved.
+  Copyright (C) 2020, PMheart. All rights reserved.
+
+  All rights reserved.
+
+  This program and the accompanying materials
+  are licensed and made available under the terms and conditions of the BSD License
+  which accompanies this distribution.  The full text of the license may be found at
+  http://opensource.org/licenses/bsd-license.php
+
+  THE PROGRAM IS DISTRIBUTED UNDER THE BSD LICENSE ON AN "AS IS" BASIS,
+  WITHOUT WARRANTIES OR REPRESENTATIONS OF ANY KIND, EITHER EXPRESS OR IMPLIED.
+**/
+
+#include "ocvalidate.h"
+
+UINT32
+CheckACPI (
+  IN  OC_GLOBAL_CONFIG  *Config
+  )
+{
+  UINT32          ErrorCount;
+  UINT32          Index;
+  OC_ACPI_CONFIG  *UserAcpi;
+  CONST CHAR8     *Path;
+  CONST CHAR8     *Comment;
+  CONST UINT8     *Find;
+  UINT32          FindSize;
+  CONST UINT8     *Replace;
+  UINT32          ReplaceSize;
+  CONST UINT8     *Mask;
+  UINT32          MaskSize;
+  CONST UINT8     *ReplaceMask;
+  UINT32          ReplaceMaskSize;
+  BOOLEAN         HasCustomDSDT;
+
+  DEBUG ((DEBUG_VERBOSE, "config loaded into ACPI checker!\n"));
+
+  ErrorCount    = 0;
+  UserAcpi      = &Config->Acpi;
+  HasCustomDSDT = FALSE;
+
+  for (Index = 0; Index < UserAcpi->Add.Count; ++Index) {
+    Path         = OC_BLOB_GET (&UserAcpi->Add.Values[Index]->Path);
+    Comment      = OC_BLOB_GET (&UserAcpi->Add.Values[Index]->Comment);
+
+    //
+    // Sanitise strings.
+    //
+    if (!AsciiFileSystemPathIsLegal (Path)) {
+      DEBUG ((DEBUG_WARN, "ACPI->Add[%u]->Path contains illegal character!\n", Index));
+      ++ErrorCount;
+    }
+    if (!AsciiCommentIsLegal (Comment)) {
+      DEBUG ((DEBUG_WARN, "ACPI->Add[%u]->Comment contains illegal character!\n", Index));
+      ++ErrorCount;
+    }
+
+    if (AsciiFileNameHasSuffix (Path, "dsl")) {
+      DEBUG ((DEBUG_WARN, "ACPI->Add[%u]->Path has .dsl suffix!\n", Index));
+      ++ErrorCount;
+    }
+
+    if (OcAsciiStriStr (Path, "DSDT") != NULL && UserAcpi->Add.Values[Index]->Enabled) {
+      HasCustomDSDT = TRUE;
+    }
+  }
+
+  for (Index = 0; Index < UserAcpi->Delete.Count; ++Index) {
+    Comment = OC_BLOB_GET (&UserAcpi->Delete.Values[Index]->Comment);
+
+    //
+    // Sanitise strings.
+    //
+    if (!AsciiCommentIsLegal (Comment)) {
+      DEBUG ((DEBUG_WARN, "ACPI->Delete[%u]->Comment contains illegal character!\n", Index));
+      ++ErrorCount;
+    }
+
+    //
+    // Size of OemTableId and TableSignature cannot be checked,
+    // as serialisation kills it.
+    //
+  }
+
+  for (Index = 0; Index < UserAcpi->Patch.Count; ++Index) {
+    Comment         = OC_BLOB_GET (&UserAcpi->Patch.Values[Index]->Comment);
+    Find            = OC_BLOB_GET (&UserAcpi->Patch.Values[Index]->Find);
+    FindSize        = UserAcpi->Patch.Values[Index]->Find.Size;
+    Replace         = OC_BLOB_GET (&UserAcpi->Patch.Values[Index]->Replace);
+    ReplaceSize     = UserAcpi->Patch.Values[Index]->Replace.Size;
+    Mask            = OC_BLOB_GET (&UserAcpi->Patch.Values[Index]->Mask);
+    MaskSize        = UserAcpi->Patch.Values[Index]->Mask.Size;
+    ReplaceMask     = OC_BLOB_GET (&UserAcpi->Patch.Values[Index]->ReplaceMask);
+    ReplaceMaskSize = UserAcpi->Patch.Values[Index]->ReplaceMask.Size;
+
+    //
+    // Sanitise strings.
+    //
+    if (!AsciiCommentIsLegal (Comment)) {
+      DEBUG ((DEBUG_WARN, "ACPI->Patch[%u]->Comment contains illegal character!\n", Index));
+      ++ErrorCount;
+    }
+
+    //
+    // Size of OemTableId and TableSignature cannot be checked,
+    // as serialisation kills it.
+    //
+
+    //
+    // Checks for size.
+    //
+    if (FindSize != ReplaceSize) {
+      DEBUG ((
+        DEBUG_WARN,
+        "ACPI->Patch[%u] has different Find and Replace size (%u vs %u)!\n",
+        Index,
+        FindSize,
+        ReplaceSize
+        ));
+      ++ErrorCount;
+    }
+    if (MaskSize > 0) {
+      if (MaskSize != FindSize) {
+        DEBUG ((
+          DEBUG_WARN,
+          "ACPI->Patch[%u] has Mask set but its size is different from Find (%u vs %u)!\n",
+          Index,
+          MaskSize,
+          FindSize
+          ));
+        ++ErrorCount;
+      } else {
+        if (!DataHasProperMasking (Find, Mask, FindSize)) {
+          DEBUG ((DEBUG_WARN, "ACPI->Patch[%u]->Find requires Mask to be active for corresponding bits!\n", Index));
+          ++ErrorCount;
+        }
+      }
+    }
+    if (ReplaceMaskSize > 0) {
+      if (ReplaceMaskSize != ReplaceSize) {
+        DEBUG ((
+          DEBUG_WARN,
+          "ACPI->Patch[%u] has ReplaceMask set but its size is different from Replace (%u vs %u)!\n",
+          Index,
+          ReplaceMaskSize,
+          ReplaceSize
+          ));
+        ++ErrorCount;
+      } else {
+        if (!DataHasProperMasking (Replace, ReplaceMask, ReplaceSize)) {
+          DEBUG ((DEBUG_WARN, "ACPI->Patch[%u]->Replace requires ReplaceMask to be active for corresponding bits!\n", Index));
+          ++ErrorCount;
+        }
+      }
+    }
+  }
+
+  //
+  // Check for RebaseRegions when using customised DSDT.
+  //
+  if (HasCustomDSDT && !UserAcpi->Quirks.RebaseRegions) {
+    DEBUG ((DEBUG_WARN, "ACPI->Quirks->RebaseRegions is not enabled when customised DSDT table is in use!\n"));
+    ++ErrorCount;
+  }
+
+  return ReportError (__func__, ErrorCount);
+}
